+++ support.py
@@ -1,6 +1,6 @@
 from aiogram import Bot, types
 from aiogram.dispatcher import Dispatcher
-from data_provider import get_go_to_menu_kb, get_bot_token, get_support_kb
+from data_provider import get_go_to_menu_kb, get_bot_token, get_support_kb, is_can_be_deleted
 
 bot = Bot(token=get_bot_token())
 dp = Dispatcher(bot)
@@ -8,7 +8,8 @@ dp = Dispatcher(bot)
 
 # support
 async def support(callback_query: types.CallbackQuery):
-    await callback_query.message.delete()
+    if is_can_be_deleted(callback_query.message.date):
+        await callback_query.message.delete()
 
     await callback_query.message.answer_photo(open("resources/pictures/support.png", "rb"),
                                               caption="Пожалуйста, выберите тему обращения ❤",
